working map function, changed long to int for ROS compat

+++ src/rcProjPkg/src/joystickProcessing.cpp
@@ -8,7 +8,7 @@ SensorProcessing::SensorProcessing(ros::NodeHandle node_handle) : node_handle(no
 }
 
 // function maps the potentiometer values ranging 0-1023 (10bit arduino ADC resolution) to a desired PWM range for RPI to handle 0-100
-long map(long x, long in_min, long in_max, long out_min, long out_max) {
+int map(int x, int in_min, int in_max, int out_min, int out_max) {
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
